Requested changes (Removed prvoiderType enum param and renamed some domains)
