+++ moove/application/src/test/groovy/io/charlescd/moove/application/metric/impl/ProviderConnectionStatusConfigurationInteractorImplUnitTest.groovy
@@ -19,46 +19,33 @@
 
 package io.charlescd.moove.application.metric.impl
 
+import io.charlescd.moove.application.configuration.impl.ProviderConnectionStatusInteractorImpl
 import io.charlescd.moove.domain.MetricConfiguration
-import io.charlescd.moove.domain.User
-import io.charlescd.moove.domain.repository.MetricConfigurationRepository
 import io.charlescd.moove.metrics.connector.MetricServiceFactory
 import io.charlescd.moove.metrics.connector.prometheus.PrometheusConnectionStatusResponse
 import io.charlescd.moove.metrics.connector.prometheus.PrometheusService
 import spock.lang.Specification
 
-import java.time.LocalDateTime
-
-class WorkspaceProviderConnectionStatusInteractorImplUnitTest extends Specification {
+class ProviderConnectionStatusConfigurationInteractorImplUnitTest extends Specification {
 
     def serviceFactory = Mock(MetricServiceFactory)
     def providerService = Mock(PrometheusService)
-    def metricConfigurationRepository = Mock(MetricConfigurationRepository)
-    def verifyProviderConnectionInteractorImpl = new VerifyWorkspaceProviderConnectionInteractorImpl(serviceFactory, metricConfigurationRepository)
+    def verifyProviderConnectionInteractorImpl = new ProviderConnectionStatusInteractorImpl(serviceFactory)
 
-    def workspaceId = "workspace-id"
-    def providerId = "provider-id"
-    def url = "http://localhost:9090"
-
-    def author = new User("81861b6f-2b6e-44a1-a745-83e298a550c9", "John Doe", "email@gmail.com",
-            "https://www.photos.com/johndoe", [], false, LocalDateTime.now())
-
-    def metricConfiguration = new MetricConfiguration("provider-id", MetricConfiguration.ProviderEnum.PROMETHEUS, url,
-            LocalDateTime.now(), workspaceId, author)
 
     def 'should return success when provider readiness is ok'() {
         given:
+        def provider = "http://localhost:9090"
         def providerType = MetricConfiguration.ProviderEnum.PROMETHEUS
         def prometheusReadinessResponse = new PrometheusConnectionStatusResponse(
                 "SUCCESS", 200, "Prometheus is Ready")
 
         when:
-        def result = verifyProviderConnectionInteractorImpl.execute(workspaceId, providerId, providerType)
+        def result = verifyProviderConnectionInteractorImpl.execute(provider, providerType)
 
         then:
         1 * serviceFactory.getConnector(providerType) >> providerService
-        1 * metricConfigurationRepository.find(providerId, workspaceId) >> Optional.of(metricConfiguration)
-        1 * providerService.readinessCheck(url) >> prometheusReadinessResponse
+        1 * providerService.readinessCheck(provider) >> prometheusReadinessResponse
         0 * _
 
         result != null
@@ -67,18 +54,17 @@ class WorkspaceProviderConnectionStatusInteractorImplUnitTest extends Specificat
 
     def 'should return failed when provider readiness is unavailable'() {
         given:
+        def provider = "http://localhost:9090"
         def providerType = MetricConfiguration.ProviderEnum.PROMETHEUS
         def prometheusReadinessResponse = new PrometheusConnectionStatusResponse(
                 "FAILED", 500, "Prometheus is not Ready")
 
         when:
-        def result = verifyProviderConnectionInteractorImpl.execute(workspaceId, providerId, providerType)
+        def result = verifyProviderConnectionInteractorImpl.execute(provider, providerType)
 
         then:
         1 * serviceFactory.getConnector(providerType) >> providerService
-        1 * metricConfigurationRepository.find(providerId, workspaceId) >> Optional.of(metricConfiguration)
-        1 * providerService.readinessCheck(url) >> prometheusReadinessResponse
-        0 * _
+        1 * providerService.readinessCheck(provider) >> { prometheusReadinessResponse }
 
         result != null
         result.status == "FAILED"
